CSHARP-2964: Code review.
