+++ tests/MongoDB.Driver.Core.TestHelpers/CoreTestConfiguration.cs
@@ -81,7 +81,12 @@ namespace MongoDB.Driver
             get
             {
                 var server = __cluster.Value.SelectServer(WritableServerSelector.Instance, CancellationToken.None);
-                return Ensure.IsNotNull(server.Description.Version, nameof(server.Description.Version));
+                var version = server.Description.Version;
+                if (version == null)
+                {
+                    throw new InvalidOperationException("ServerDescription.Version is unexpectedly null.");
+                }
+                return version;
             }
         }
